Reference global.XMLHttpRequest directly
- check for global.XMLHttpRequest directly instead of what
global.XMLHttRequest referenced when the library was first parsed
- reference global.XMLHttpRequest constructor instead of the constructor
that was dereferenced when the library was first parsed
- reference global.XMLHttpRequest status properties directly instead of
the properties of the prototype of the constructor that was dereferenced
when the library was first parsed

This will allow mocking libraries that rely on being able to replace
the reference to global.XMLHttpRequest with a mock XMLHttpRequest object
that can be programmatically manipulated.

Such libraries include the jasmine-ajax plugin. I'm sure there are
others.

It's a little more typing, but it fixes a big bug.